use the dnf-helper to do installs as a speed boost
eliminates the need to flush the cache entirely after doing install
actions.

speed boost on a set of 14 chef dnf_package resources (not
multipackage) is:

with the patch:
  0.030000   0.060000  18.070000 ( 19.413037)

without the patch:
  0.060000   0.130000  24.470000 ( 27.511077)

based on those numbers i think the major problem is that dnf is slow
and that multipackage installs is a better approach to speeding them
up.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>